Se agregan las listas ingredientes y empleados

+++ src/model/Ingredient.java
@@ -4,9 +4,9 @@ public class Ingredient {
 	
 	private String name;
 	private MEASUREMENT_TYPE measurement;
-	private int amount;
+	private double amount;
 	
-	public Ingredient(String name, MEASUREMENT_TYPE measurement, int amount) {
+	public Ingredient(String name, MEASUREMENT_TYPE measurement, double amount) {
 		this.name = name;
 		this.measurement = measurement;
 		this.amount = amount;
@@ -23,7 +23,7 @@ public class Ingredient {
 	public void setMeasurement(MEASUREMENT_TYPE measurement) {
 		this.measurement = measurement;
 	}
-	public int getAmount() {
+	public double getAmount() {
 		return amount;
 	}
 	public void setAmount(int amount) {
